quick and dirty fix

+++ pulseaudio-system/install.sh
@@ -22,8 +22,14 @@ ETCDIR=/etc
 if [ "$DISTRO" = "FreeBSD" ]; then
 	ETCDIR=/usr/local/etc
 fi	
-cp etc-profile/* $ETCDIR/profile.d/
-cp etc-pulse/*   $ETCDIR/pulse/
+if [ -f /etc/environment ]; then
+	if [ $(grep -c PULSE_SERVER /etc/environment) -eq 0 ]; then
+		echo "PULSE_SERVER=unix:/tmp/pulse-server" >> /etc/environment
+	fi
+else
+	cp etc-profile/* $ETCDIR/profile.d/
+	cp etc-pulse/*   $ETCDIR/pulse/
+fi
 
 # pulse user
 #username="pulse"
